dmaengine: rcar-dmac: Compute maximum chunk size at runtime
The DMAC expressed DMA request lengths as a number of transfers, limited
to 16777216 by the hardware. As a single transfer transfers a
configurable number of bytes (from 1 to 64), the chunk size limit isn't
fixed and must be computed at runtime by multiplying the transfers
number limit by the transfer size.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>